Rename "NotFound" to "Unhandled"
When a router's shared by the front- and back- ends, these concepts aren't the
same. On the browser, the resource a path corresponds to could be not found, but
the UI may still update to reflect that (in which case, the server should return
a 404 with the UI in the body). Alternatively, if the UI is not updated (i.e.
the React app is not responsible for generating the 404 view), the server app
should be allowed to handle it in some other manner.