Clarify responsibilities of invokeHandlers
This rolls back ee3dbcd and removes all of its responsibility for examining or
manipulating the response. It now just calls middleware and invokes a callback
when they're done.

Relaying middleware errors to the response is now done in `dispatch` in the
callback passed to `invokeHandlers`. This means that the response object is now
the canonical source of errors and they are simply relayed to the `dispatch`
callback.

Examining `res.ended` was also removed from `invokeHandlers` because middleware
that end a response should not call `next()`.

So much cleaner. Closes GH-23, I think.